--- src/components/Profile.js
@@ -1,23 +1,0 @@
-import React, { Component } from 'react';
-import { View, Text } from 'react-native';
-
-class Profile extends Component {
-    static navigationOptions = {
-        title: 'Profile',
-    }
-
-    render() {
-        return (
-            <View
-            style={{
-                flex: 1,
-                justifyContent: 'center',
-                alignItems: 'center'
-              }}>
-                <Text>Profile</Text>
-            </View>
-        );
-    }
-}
-
-export default Profile;
