+++ src/components/screens/List.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import { View, Text } from 'react-native';
+
+class List extends Component {
+  static navigationOptions = {
+    title: 'List'
+  };
+
+  render() {
+    return (
+      <View>
+        <Text>List</Text>
+      </View>
+    );
+  }
+}
+
+export default List;
