+++ src/components/screens/Map.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import { View, Text } from 'react-native';
+
+class Map extends Component {
+  static navigationOptions = {
+    title: 'Map'
+  };
+
+  render() {
+    return (
+      <View>
+        <Text>Map</Text>
+      </View>
+    );
+  }
+}
+
+export default Map;
