+++ src/components/screens/More.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import { View, Text } from 'react-native';
+
+class More extends Component {
+  static navigationOptions = {
+    title: 'More'
+  };
+
+  render() {
+    return (
+        <View>
+            <Text>More</Text>
+        </View>
+    );
+  }
+}
+
+export default More;
